+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 
 # 2.19.3
 - Fixed index error related to Totem sigils
+- Fixed index error related to opponent sniper targeting
 
 # 2.19.2
 - Fixed activated abilities not being interactable in Act 3